[DEC-2081] Audit app/ante/sigverify.go (#323)
* [DEC-2081] Audit app/ante/sigverify.go

Note that I dropped tests that just ran over the SDKs gas cost and set pubkey decorators as these we don't need to maintain. This allowed me to remove some additional methods.

I also confirmed that the sigverification tests that the SDK has hadn't changed and we cover the new functionality with our own tests that we added.