Impose Broker Fees on Order Matching
When an order is matched, the current revenue sharing model assumes
that the protocol fees that are charged to the user are going to
be split amongst the protocol and affiliates. If we want to impose
a broker fee (top up etc) on this, we need to modify the collateral
checks that happen when an order is placed as well as remove this
fee from the subaccounts when the trade is matched.