feat(api): don't strike users from /geoblock*, raise 403
If a connection is restricted, raise a 403 geo-blocking error
immediately and not process request.

If an address is CLOSE_ONLY, connection does not upgrade
status.

If an address is BLOCKED, an ok connection returns COMPLIANT
and a restricted connection moves the address to CLOSE_ONLY.

If an address is in FIRST_STRIKE, or FIRST_STRIKE_CLOSE_ONLY
and the connection is ok, return COMPLIANT. If the connection
is restricted, move address to CLOSE_ONLY.

If VALID_SURVEY and connection is ok, return COMPLIANT.

If INVALID_SURVEY or connection is restricted, move address
to CLOSE_ONLY.